docServer/requirements.txt
psycopg2==2.6
hiredis==0.2.0
Django==1.7.5
django-bootstrap-form==3.1
django-bootstrap-pagination==1.5.1
django-form-utils==1.0.2
django-storages==1.1.8
django-user-accounts==1.0.1
gunicorn==19.2.1
pinax-theme-bootstrap==5.7.2
django-require==1.0.7
fabric==1.8
djangorestframework==3.1.1
djangorestframework-csv==1.3.3
markdown==2.6.1
django-filter==0.9.2
ajaxuploader==0.3.8
django-behave==0.1.3
pyvirtualdisplay==0.1.5
wsgi_intercept==0.9.1
django-phonenumber-field==0.7.2
pytz==2014.10
django-appconf==1.0.1
django-enumfield==1.2.1
django-session-security==2.2.3
ipdb==0.8.1
django-redis-cache==0.13.1
django-report-builder==3.1.8
#python-dateutil==2.4.0
#python-mimeparse==0.1.4
#six==1.9.0
#pycrypto